+++ src/components/Landing/Landing.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import {
   Box,
   Breadcrumbs,
-  Footer,
   Heading,
   Hero,
   Link,
@@ -107,9 +106,7 @@ const LandingPage: React.FC<{ collections?: Opds2Feed }> = ({
           </Box>
         </TemplateContentPrimary>
       </TemplateContent>
-      <TemplateFooter>
-        <Footer />
-      </TemplateFooter>
+      <TemplateFooter />
     </Template>
   );
 };