+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGE LOG
 
 ## [Pre-release]
+- Replace dgx-header-component and DS Footer with Header/Footer components from nypl-header-app 
 - Refactored Playwright/Cucumber tests to use new click action function
 - Refactored Playwright/Cucumber tests to use new dropdown action function
 
@@ -11,8 +12,6 @@
 - Add Collection page
 - Add collection placeholder image
 - Update hero image to update based on screen size
-- Chore: fresh package-lock.json install and fix install errors on mac with M1 chip
-- Add Playwright/Cucumber tests to support new Recently Added Collections feature on landing page
 - Chore: fresh package-lock.json install and fix install errors on mac with M1 chip
 - Add Playwright/Cucumber tests to support new Recently Added Collections feature on landing page
 - Refactored Playwright/Cucumber tests to use new page visit actions function